Bypassed shmem support for non-persistent tools under AFL++
Non-persistent tools that are linked against the AFL++ driver indicate
to the fuzzer that they use shmem, but in fact, they use stdin or a file
name argument. This patch bypasses the AFL++ check on shmem. It still
uses shmem for targets that support it, but it also checks if the target
has requested a file name argument.

This is a temporary fix until a better method is found to compile
non-persistent tools generically.